Rename device clock class files to match their class names
