+++ lib/bus/mac/mac.cpp
@@ -216,7 +216,7 @@ macBus MAC; // global smartport bus variable
 #include "../device/iwm/disk2.h"
 #include "../device/iwm/fuji.h"
 #include "../device/iwm/cpm.h"
-#include "../device/iwm/clock.h"
+#include "../device/iwm/iwmClock.h"
 
 /******************************************************************************
 Based on:
