+++ lib/device/drivewire/drivewireClock.cpp
@@ -1,6 +1,6 @@
 #ifdef BUILD_COCO
 
-#include "clock.h"
+#include "drivewireClock.h"
 
 #include <algorithm>
 #include <optional>
